education/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
blis==1.0.1
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools==5.5.0
catalogue==2.0.10
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.8
decorator==4.4.2
Django==5.1.2
executing==2.1.0
faiss-cpu==1.9.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fsspec==2024.10.0
gmpy2 @ file:///C:/ci_311/gmpy2_1677743390134/work
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
httplib2==0.22.0
huggingface-hub==0.26.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
ipython==8.29.0
jedi==0.19.1
Jinja2 @ file:///C:/b/abs_92fccttino/croot/jinja2_1716993447201/work
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/b/abs_ecfdqh67b_/croot/markupsafe_1704206030535/work
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/b/abs_f55mv94vyg/croot/mkl_fft_1725370278455/work
mkl_random @ file:///C:/b/abs_21ydbzdu8d/croot/mkl_random_1725370276095/work
moviepy==1.0.3
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
murmurhash==1.0.10
mutagen==1.47.0
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp311-cp311-win_amd64.whl#sha256=46eb5850478cda64bda6d2f2d3d6ad8ceeb7cc168b83532e4f9dbbcb42e22ab2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow @ file:///C:/b/abs_32o8er3uqp/croot/pillow_1721059447598/work
preshed==3.0.9
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.12
pyparsing==3.2.0
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
regex==2024.9.11
reportlab==4.2.5
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
rich==13.9.3
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.1
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.0
typer==0.12.5
typing_extensions @ file:///C:/b/abs_0as9mdbkfl/croot/typing_extensions_1715268906610/work
tzdata==2024.2
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==13.1
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
wrapt==1.16.0
yt-dlp==2024.10.22